feat(query) Support no downtime query capabilities in case of time split partitions
